refactor: home ozone/aerosol models in atmospheric_models
The ozone profile tables (OzZeta/OzDepth/OzDsum), aerosol OD table (aOD55),
their derived exact layer-walk tables, ozone_losses/ozone_rate, and the
analytic ozone_column/aerosol_column/refractive_index functions are all
atmosphere physics with no coupling to the Cherenkov orchestration -- the
call graph clusters them apart from CphotAng.run. Move them (verbatim) to
atmospheric_models.py as module-level functions and constants; CphotAng
keeps thin ozone_losses/ozone_rate delegates and re-exports for
compatibility. Physics goldens unchanged (bit-identical).
